requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.1.1
Flask-Mail==0.9.1
Flask-Cors==3.1.1
